Readable stream socket now fulfills with an empty string on EOF.
The promise returned by read() on a readable stream socket is now fulfilled with an empty string if the connection is closed instead of being rejected with an EofException. Subsquent calls to read will still reject as before. This behavior allows reading from a stream socket in a loop in a coroutine without having to use a try/catch.