Fix bug related to generators that never yield.
If a generator is written in such a way that it may never yield, the coroutine is rejected even though the generator had not been run and could have yielded values under different circumstances. This patch alters Coroutine to resolve with null in this scenario instead of rejecting the coroutine.